moving sidebar and responsiveDrawer into the UI folder and updating import statements
